+++ test/Microsoft.ApplicationInsights.AspNetCore.Tests/Helpers/CommonMocks.cs
@@ -6,15 +6,16 @@
 
     public static class CommonMocks
     {
+        public const string InstrumentationKey = "REQUIRED";
+        public const string InstrumentationKeyHash = "0KNjBVW77H/AWpjTEcI7AP0atNgpasSkEll22AtqaVk=";
+
         public static TelemetryClient MockTelemetryClient(Action<ITelemetry> onSendCallback)
         {
-            var telemetryChannel = new FakeTelemetryChannel { OnSend = onSendCallback };
-
-            var telemetryConfiguration = new TelemetryConfiguration();
-            telemetryConfiguration.InstrumentationKey = "REQUIRED";
-            telemetryConfiguration.TelemetryChannel = telemetryChannel;
-
-            return new TelemetryClient(telemetryConfiguration);
+            return new TelemetryClient(new TelemetryConfiguration()
+            {
+                InstrumentationKey = InstrumentationKey,
+                TelemetryChannel = new FakeTelemetryChannel { OnSend = onSendCallback }
+            });
         }
     }
 }